+++ cloud/aws/process.go
@@ -15,13 +15,13 @@ func (p *AwsCloud) ProcessList(app string) ([]*pb.Process, error) {
 	return kube.ProcessList(p.kubeClient(), p.DeploymentName, app)
 }
 
-func (p *AwsCloud) ProcessRun(name string, r io.ReadWriter, command []string) error {
+func (p *AwsCloud) ProcessRun(name string, r io.ReadWriter, opts pb.ProcessRunOptions) error {
 	ns := p.DeploymentName
 	cfg, _ := getKubeClientConfig(ns)
 	envVars, _ := p.EnvironmentGet(name)
 	app, _ := p.AppGet(name)
 
-	return kube.ProcessRun(p.kubeClient(), cfg, ns, name, p.latestImage(app), command, envVars, false, r, cloud.AwsProvider)
+	return kube.ProcessRun(p.kubeClient(), cfg, ns, name, p.latestImage(app), opts, envVars, false, r, cloud.AwsProvider)
 }
 
 func (p *AwsCloud) ProcessSave(name string, structure map[string]int32) error {