+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASED] - 0000-00-00
+### Added
+- Support terminal raw mode for process exec/run cli
 
 ## [1.0.0-alpha.16] - 2018-12-21
 ### Added